server/requirements.txt
# =============================================================================
# server/requirements.txt
#
# Install with:  pip install -r requirements.txt
#
# Run server with:  uvicorn main:app --reload --port 8000
# =============================================================================

# Web framework and ASGI server
fastapi
uvicorn[standard]

# Required for FastAPI to accept multipart/form-data file uploads
python-multipart

# Loads GEMINI_API_KEY (and later MONGODB_URI) from server/.env
python-dotenv

# Gemini AI -- used in audit.py for vision, text generation, and embeddings
google-generativeai

# ElevenLabs -- used when wiring the Conversational AI agent
# The Python SDK is for server-side tool call handling if needed.
# The React SDK (@elevenlabs/react) handles the browser-side voice session.
elevenlabs

# MongoDB Atlas -- used for Battery Digital Twin storage and Vector Search
# [srv] installs the DNS seedlist dependency needed for Atlas connection strings
pymongo[srv]
